A number of small refactorings

+++ src/java/fault/BlockingExecutor.java
@@ -35,7 +35,7 @@ public class BlockingExecutor extends AbstractServiceExecutor implements Service
     }
 
     @Override
-    public <T> ResilientPromise<T> performAction(final ResilientAction<T> action, int millisTimeout) {
+    public <T> ResilientPromise<T> performAction(final ResilientAction<T> action, long millisTimeout) {
         final ResilientPromise<T> promise = new MultipleWriterResilientPromise<>();
         final Future<Void> f = service.submit(new Callable<Void>() {
             @Override
